Added a bunch of attached behaviors:
   -AttachedAdorner for the ability to declare adorners in XAML
   -SmartSizer to hide controls specified as Detail controls when there is not enough room

VisualTreeExtensions, SizeExtensions, and DictionaryExtensions to help with the APIs.

Initial display of the build map when the user clicks the build button that shows up on Hover.
This uses the selected item in the listbox for now until I can figure out how to wire up the CommandTarget properly.
Also added lame-o ShowBuildMap event to the viewmodel because I'm too lazy to do it the right way for now.